fix(ui): close composer invocation regressions
