+++ ui/src/pages/chat/components/chat-composer-skill-menu.ts
@@ -8,8 +8,11 @@ import {
   getSlashCommandDescription,
   type SlashCommandDef,
 } from "../../../lib/chat/commands.ts";
-import { paneDomId, scrollActiveMenuOptionIntoView } from "./chat-composer-dom.ts";
-import { syncComposerMenuScroll } from "./chat-composer-slash-menu.ts";
+import {
+  paneDomId,
+  scrollActiveMenuOptionIntoView,
+  syncComposerMenuScroll,
+} from "./chat-composer-dom.ts";
 
 const SKILL_MENTION_CHAR = /[-a-zA-Z0-9_:]/u;
