+++ ui/src/pages/chat/components/chat-composer-slash-menu.ts
@@ -10,7 +10,11 @@ import {
   type SlashCommandCategory,
   type SlashCommandDef,
 } from "../../../lib/chat/commands.ts";
-import { paneDomId, scrollActiveMenuOptionIntoView } from "./chat-composer-dom.ts";
+import {
+  paneDomId,
+  scrollActiveMenuOptionIntoView,
+  syncComposerMenuScroll,
+} from "./chat-composer-dom.ts";
 
 export type SlashMenuState = {
   slashMenuOpen: boolean;
@@ -285,22 +289,6 @@ function slashOptionIdSegment(value: string): string {
       .replace(/[^a-z0-9_-]+/gu, "-")
       .replace(/^-+|-+$/gu, "") || "item"
   );
-}
-
-export function syncComposerMenuScroll(element: Element | undefined): void {
-  if (!(element instanceof HTMLElement)) {
-    return;
-  }
-  const sync = () => {
-    const scrollable = element.scrollHeight > element.clientHeight + 1;
-    element.dataset.scrollable = String(scrollable);
-    element.dataset.atStart = String(!scrollable || element.scrollTop <= 1);
-    element.dataset.atEnd = String(
-      !scrollable || element.scrollTop + element.clientHeight >= element.scrollHeight - 1,
-    );
-  };
-  sync();
-  requestAnimationFrame(sync);
 }
 
 function syncComposerMenuScrollEvent(event: Event): void {
