+++ ui/src/styles/chat/progress-card.css
@@ -2,7 +2,6 @@
   --session-progress-neutral: var(--muted);
   --session-progress-marker-size: 14px;
 
-  box-sizing: border-box;
   min-width: 0;
   color: var(--text);
 }
@@ -193,7 +192,6 @@
 .session-hovercard__creator-avatar,
 .session-hovercard__creator-avatar .viewer-avatar {
   display: block;
-  box-sizing: border-box;
   width: var(--session-hovercard-context-icon-size);
   height: var(--session-hovercard-context-icon-size);
   aspect-ratio: 1;
@@ -493,7 +491,6 @@
 .session-progress-card__current-marker,
 .session-progress-card__step-marker {
   display: inline-flex;
-  box-sizing: border-box;
   width: var(--session-progress-marker-size);
   height: var(--session-progress-marker-size);
   align-items: center;
