+++ ui/src/styles/chat/text.css
@@ -166,7 +166,6 @@
 .markdown-table__viewport {
   max-width: 100%;
   overflow-x: auto;
-  scrollbar-width: thin;
 }
 
 .markdown-table--can-scroll-right .markdown-table__viewport {
@@ -300,7 +299,6 @@
     color-mix(in srgb, var(--accent) 6%, var(--secondary))
   );
   color: var(--text-strong);
-  font-family: var(--font-body);
   font-weight: 600;
 }
 
@@ -417,7 +415,6 @@
 }
 
 :is(.chat-text, .chat-thinking) :where(details) {
-  box-sizing: border-box;
   width: 100%;
   padding: var(--space-4);
   border: 1px solid var(--border);
@@ -724,7 +721,6 @@
 
 :is(.chat-text, .sidebar-markdown) .markdown-external-image {
   display: flex;
-  box-sizing: border-box;
   align-items: center;
   justify-content: space-between;
   flex-wrap: wrap;
